Extract disabling methods for view-only layouts to mixin
The inheritance for the controllers in question
is odd and needs looking at, so have removed
to mixin and included as a quick fix.

Needs testings and tests.
