+++ app/controllers/concerns/google_analytics_actions.rb
@@ -1,0 +1,19 @@
+module DisableViewOnlyActions
+  extend ActiveSupport::Concern
+
+  included do
+    before_action :disable_analytics, :disable_phase_banner, :disable_flashes
+  end
+
+  def disable_analytics
+    @disable_analytics = true
+  end
+
+  def disable_phase_banner
+    @disable_phase_banner = true
+  end
+
+  def disable_flashes
+    @disable_flashes = true
+  end
+end
