+++ app/controllers/geckoboard_api/widgets_controller.rb
@@ -1,5 +1,5 @@
 class GeckoboardApi::WidgetsController < GeckoboardApi::ApplicationController
-  before_action :disable_analytics, :disable_phase_banner, :disable_flashes
+  include DisableViewOnlyActions
 
   def claims
     respond_payload_from_class(Stats::ClaimPercentageAuthorisedGenerator)
